requirements.txt
scipy==1.4.1
matplotlib==3.1.1
plac==1.1.0
pillow==6.2.2
numpy==1.18.1
pandas==1.0.1
wandb==0.8.34
efficientnet-pytorch==0.6.3
opencv-python==3.4.11.45
open3d==0.11.0
# Original PyTorch version
#pytorch==1.3.1
#torchvision==0.4.2